+++ src/decoder/png.rs
@@ -1,3 +1,5 @@
+use std::sync::Once;
+
 use crate::{
     DecodeError, Grib2GpvUnpack, Grib2SubmessageDecoder,
     decoder::{
@@ -5,6 +7,8 @@ use crate::{
         stream::{FixedValueIterator, NBitwiseIterator},
     },
 };
+
+static PNG_NUM_BITS_ZERO_WARNING_ONCE: Once = Once::new();
 
 pub(crate) struct Png<'d>(
     pub(crate) &'d Grib2SubmessageDecoder,
@@ -21,14 +25,13 @@ impl<'d> Grib2GpvUnpack for Png<'d> {
         let Self(target, template) = self;
         super::orig_field_type_is_supported(template.orig_field_type)?;
 
-        let buf = read_image_buffer(target.sect7_payload())
-            .map_err(|e| DecodeError::from(format!("PNG decode error: {e}")))?;
-
         if template.simple.num_bits == 0 {
-            eprintln!(
-                "WARNING: nbit = 0 for PNG decoder is not tested.
-                Please report your data and help us develop the library."
-            );
+            PNG_NUM_BITS_ZERO_WARNING_ONCE.call_once(|| {
+                eprintln!(
+                    "WARNING: data with num_bits being 0 for PNG decoder is not tested.
+                    Please report your data and help us develop the library."
+                );
+            });
             let decoder = SimplePackingDecoder::ZeroLength(FixedValueIterator::new(
                 template.simple.zero_bit_reference_value(),
                 target.num_encoded_points(),
@@ -36,13 +39,18 @@ impl<'d> Grib2GpvUnpack for Png<'d> {
             return Ok(decoder);
         };
 
-        if template.simple.num_bits != 16 {
-            eprintln!(
-                "WARNING: nbit != 16 for PNG decoder is not tested.
-                Please report your data and help us develop the library."
-            );
-        }
+        // Valid values for `num_bits` can be:
+        //
+        // - 1, 2, 4, 8, or 16 : Treat as greyscale image
+        // - 24 : Treat as RGB colour image (each component having 8-bit depth)
+        // - 32 : Treat as RGB w/ alpha sample colour image (each component having 8-bit
+        //   depth)
+        //
+        // Since we have verified the decoding process through testing for cases where
+        // `num_bits` is 8, 16, and 24, we believe other cases should also be fine.
 
+        let buf = read_image_buffer(target.sect7_payload())
+            .map_err(|e| DecodeError::from(format!("PNG decode error: {e}")))?;
         let iter = NBitwiseIterator::new(buf, usize::from(template.simple.num_bits));
         let iter = NonZeroSimplePackingDecoder::new(iter, &template.simple);
         let iter = SimplePackingDecoder::NonZeroLength(iter);