+++ src/grid/latlon.rs
@@ -1,116 +1,28 @@
-use super::{
-    GridPointIndexIterator, ScanningMode,
-    helpers::{RegularGridIterator, evenly_spaced_degrees, evenly_spaced_longitudes},
-};
+use super::helpers::{RegularGridIterator, evenly_spaced_degrees, evenly_spaced_longitudes};
 use crate::{
-    error::GribError,
-    helpers::{GribInt, read_as},
+    GridPointIndex, LatLons, def::grib2::template::param_set, error::GribError, grid::AngleUnit,
 };
 
-#[derive(Debug, PartialEq, Eq)]
-pub struct LatLonGridDefinition {
-    pub ni: u32,
-    pub nj: u32,
-    pub first_point_lat: i32,
-    pub first_point_lon: i32,
-    pub last_point_lat: i32,
-    pub last_point_lon: i32,
-    pub scanning_mode: ScanningMode,
-}
-
-impl LatLonGridDefinition {
-    /// Returns the shape of the grid, i.e. a tuple of the number of grids in
-    /// the i and j directions.
-    ///
-    /// Examples
-    ///
-    /// ```
-    /// let def = grib::LatLonGridDefinition {
-    ///     ni: 2,
-    ///     nj: 3,
-    ///     first_point_lat: 0,
-    ///     first_point_lon: 0,
-    ///     last_point_lat: 2_000_000,
-    ///     last_point_lon: 1_000_000,
-    ///     scanning_mode: grib::ScanningMode(0b01000000),
-    /// };
-    /// let shape = def.grid_shape();
-    /// assert_eq!(shape, (2, 3));
-    /// ```
-    pub fn grid_shape(&self) -> (usize, usize) {
-        (self.ni as usize, self.nj as usize)
-    }
-
-    /// Returns the grid type.
-    pub fn short_name(&self) -> &'static str {
+impl crate::GridShortName for param_set::LatLonGrid {
+    fn short_name(&self) -> &'static str {
         "regular_ll"
     }
+}
 
-    /// Returns an iterator over `(i, j)` of grid points.
-    ///
-    /// Note that this is a low-level API and it is not checked that the number
-    /// of iterator iterations is consistent with the number of grid points
-    /// defined in the data.
-    ///
-    /// Examples
-    ///
-    /// ```
-    /// let def = grib::LatLonGridDefinition {
-    ///     ni: 2,
-    ///     nj: 3,
-    ///     first_point_lat: 0,
-    ///     first_point_lon: 0,
-    ///     last_point_lat: 2_000_000,
-    ///     last_point_lon: 1_000_000,
-    ///     scanning_mode: grib::ScanningMode(0b01000000),
-    /// };
-    /// let ij = def.ij();
-    /// assert!(ij.is_ok());
-    ///
-    /// let mut ij = ij.unwrap();
-    /// assert_eq!(ij.next(), Some((0, 0)));
-    /// assert_eq!(ij.next(), Some((1, 0)));
-    /// assert_eq!(ij.next(), Some((0, 1)));
-    /// ```
-    pub fn ij(&self) -> Result<GridPointIndexIterator, GribError> {
-        if self.scanning_mode.has_unsupported_flags() {
-            let ScanningMode(mode) = self.scanning_mode;
-            return Err(GribError::NotSupported(format!("scanning mode {mode}")));
-        }
-
-        let iter =
-            GridPointIndexIterator::new(self.ni as usize, self.nj as usize, self.scanning_mode);
-        Ok(iter)
+impl GridPointIndex for param_set::LatLonGrid {
+    fn grid_shape(&self) -> (usize, usize) {
+        (self.grid.ni as usize, self.grid.nj as usize)
     }
 
-    /// Returns an iterator over latitudes and longitudes of grid points in
-    /// degrees.
-    ///
-    /// Note that this is a low-level API and it is not checked that the number
-    /// of iterator iterations is consistent with the number of grid points
-    /// defined in the data.
-    ///
-    /// Examples
-    ///
-    /// ```
-    /// let def = grib::LatLonGridDefinition {
-    ///     ni: 2,
-    ///     nj: 3,
-    ///     first_point_lat: 0,
-    ///     first_point_lon: 0,
-    ///     last_point_lat: 2_000_000,
-    ///     last_point_lon: 1_000_000,
-    ///     scanning_mode: grib::ScanningMode(0b01000000),
-    /// };
-    /// let latlons = def.latlons();
-    /// assert!(latlons.is_ok());
-    ///
-    /// let mut latlons = latlons.unwrap();
-    /// assert_eq!(latlons.next(), Some((0.0, 0.0)));
-    /// assert_eq!(latlons.next(), Some((0.0, 1.0)));
-    /// assert_eq!(latlons.next(), Some((1.0, 0.0)));
-    /// ```
-    pub fn latlons(&self) -> Result<RegularGridIterator, GribError> {
+    fn scanning_mode(&self) -> &param_set::ScanningMode {
+        &self.scanning_mode
+    }
+}
+
+impl LatLons for param_set::LatLonGrid {
+    type Iter<'a> = RegularGridIterator;
+
+    fn latlons_unchecked<'a>(&'a self) -> Result<Self::Iter<'a>, GribError> {
         if !self.is_consistent_for_j() {
             return Err(GribError::InvalidValueError(
                 "Latitudes for first/last grid points are not consistent with scanning mode"
@@ -119,73 +31,78 @@ impl LatLonGridDefinition {
         }
 
         let ij = self.ij()?;
+        let angle_units = self.angle_unit() as f32;
         let lat = evenly_spaced_degrees(
-            self.first_point_lat as f32,
-            self.last_point_lat as f32,
-            (self.nj - 1) as usize,
+            self.grid.first_point_lat as f32,
+            self.grid.last_point_lat as f32,
+            (self.grid.nj - 1) as usize,
+            angle_units,
         );
         let lon = evenly_spaced_longitudes(
-            self.first_point_lon,
-            self.last_point_lon,
-            (self.ni - 1) as usize,
+            self.grid.first_point_lon,
+            self.grid.last_point_lon,
+            (self.grid.ni - 1) as usize,
+            angle_units,
             self.scanning_mode,
         );
 
         let iter = RegularGridIterator::new(lat, lon, ij);
         Ok(iter)
     }
+}
 
-    pub(crate) fn is_consistent_for_j(&self) -> bool {
-        let lat_diff = self.last_point_lat - self.first_point_lat;
-        !((lat_diff > 0) ^ self.scanning_mode.scans_positively_for_j())
+impl AngleUnit for param_set::LatLonGrid {
+    fn angle_unit(&self) -> f64 {
+        self.grid.angle_unit()
     }
+}
 
-    pub(crate) fn from_buf(buf: &[u8]) -> Self {
-        let ni = read_as!(u32, buf, 0);
-        let nj = read_as!(u32, buf, 4);
-        let first_point_lat = read_as!(u32, buf, 16).as_grib_int();
-        let first_point_lon = read_as!(u32, buf, 20).as_grib_int();
-        let last_point_lat = read_as!(u32, buf, 25).as_grib_int();
-        let last_point_lon = read_as!(u32, buf, 29).as_grib_int();
-        let scanning_mode = read_as!(u8, buf, 41);
-        Self {
-            ni,
-            nj,
-            first_point_lat,
-            first_point_lon,
-            last_point_lat,
-            last_point_lon,
-            scanning_mode: ScanningMode(scanning_mode),
-        }
+impl param_set::LatLonGrid {
+    pub(crate) fn is_consistent_for_j(&self) -> bool {
+        let lat_diff = self.grid.last_point_lat - self.grid.first_point_lat;
+        !((lat_diff > 0) ^ self.scanning_mode.scans_positively_for_j())
     }
 }
 
 #[cfg(test)]
 mod tests {
     use super::*;
+    use crate::grid::helpers::test_helpers::assert_coord_almost_eq;
 
     macro_rules! test_lat_lon_calculation_for_inconsistent_longitude_definitions {
         ($((
             $name:ident,
             $grid:expr,
+            $scanning_mode:expr,
             $expected_head:expr,
             $expected_tail:expr
         ),)*) => ($(
             #[test]
             fn $name() {
-                let grid = $grid;
+                let grid = param_set::LatLonGrid {
+                    grid: $grid,
+                    i_direction_inc: 0xffffffff,
+                    j_direction_inc: 0xffffffff,
+                    scanning_mode: $scanning_mode,
+                };
                 let latlons = grid.latlons();
                 assert!(latlons.is_ok());
+
+                let delta = 1e-4;
 
                 let latlons = latlons.unwrap();
                 let actual = latlons.clone().take(3).collect::<Vec<_>>();
                 let expected = $expected_head;
-                assert_eq!(actual, expected);
+                for (a, e) in actual.iter().zip(expected) {
+                    assert_coord_almost_eq(*a, e, delta);
+                }
 
                 let (len, _) = latlons.size_hint();
                 let actual = latlons.skip(len - 3).collect::<Vec<_>>();
                 let expected = $expected_tail;
-                assert_eq!(actual, expected);
+                for (a, e) in actual.iter().zip(expected) {
+                    assert_coord_almost_eq(*a, e, delta);
+                }
             }
         )*);
     }
@@ -193,61 +110,73 @@ mod tests {
     test_lat_lon_calculation_for_inconsistent_longitude_definitions! {
         (
             lat_lon_calculation_for_increasing_longitudes_and_positive_direction_scan,
-            LatLonGridDefinition {
+            param_set::Grid {
                 ni: 1500,
                 nj: 751,
+                initial_production_domain_basic_angle: 0,
+                basic_angle_subdivisions: 0xffffffff,
                 first_point_lat: -90000000,
                 first_point_lon: 0,
+                resolution_and_component_flags: param_set::ResolutionAndComponentFlags(0b00110000),
                 last_point_lat: 90000000,
                 last_point_lon: 359760000,
-                scanning_mode: ScanningMode(0b01000000),
             },
+            param_set::ScanningMode(0b01000000),
             vec![(-90.0, 0.0), (-90.0, 0.24), (-90.0, 0.48)],
-            vec![(90.0, 359.28), (90.0, 359.52), (90.0, 359.76)]
+            vec![(90.0, -0.72), (90.0, -0.48), (90.0, -0.24)]
         ),
         (
             // grid point definition extracted from
             // testdata/CMC_glb_TMP_ISBL_1_latlon.24x.24_2021051800_P000.grib2
             lat_lon_calculation_for_decreasing_longitudes_and_positive_direction_scan,
-            LatLonGridDefinition {
+            param_set::Grid {
                 ni: 1500,
                 nj: 751,
+                initial_production_domain_basic_angle: 0,
+                basic_angle_subdivisions: 0xffffffff,
                 first_point_lat: -90000000,
                 first_point_lon: 180000000,
+                resolution_and_component_flags: param_set::ResolutionAndComponentFlags(0b00110000),
                 last_point_lat: 90000000,
                 last_point_lon: 179760000,
-                scanning_mode: ScanningMode(0b01000000),
             },
-            vec![(-90.0, 180.0), (-90.0, 180.24), (-90.0, 180.48)],
+            param_set::ScanningMode(0b01000000),
+            vec![(-90.0, -180.0), (-90.0, -179.76), (-90.0, -179.52)],
             vec![(90.0, 179.28003), (90.0, 179.52002), (90.0, 179.76001)]
         ),
         (
             lat_lon_calculation_for_decreasing_longitudes_and_negative_direction_scan,
-            LatLonGridDefinition {
+            param_set::Grid {
                 ni: 1500,
                 nj: 751,
+                initial_production_domain_basic_angle: 0,
+                basic_angle_subdivisions: 0xffffffff,
                 first_point_lat: -90000000,
                 first_point_lon: 359760000,
+                resolution_and_component_flags: param_set::ResolutionAndComponentFlags(0b00110000),
                 last_point_lat: 90000000,
                 last_point_lon: 0,
-                scanning_mode: ScanningMode(0b11000000),
             },
-            vec![(-90.0, 359.76), (-90.0, 359.52), (-90.0, 359.28)],
+            param_set::ScanningMode(0b11000000),
+            vec![(-90.0, -0.24), (-90.0, -0.48), (-90.0, -0.72)],
             vec![(90.0, 0.48), (90.0, 0.24), (90.0, 0.0)]
         ),
         (
             lat_lon_calculation_for_increasing_longitudes_and_negative_direction_scan,
-            LatLonGridDefinition {
+            param_set::Grid {
                 ni: 1500,
                 nj: 751,
+                initial_production_domain_basic_angle: 0,
+                basic_angle_subdivisions: 0xffffffff,
                 first_point_lat: -90000000,
                 first_point_lon: 179760000,
+                resolution_and_component_flags: param_set::ResolutionAndComponentFlags(0b00110000),
                 last_point_lat: 90000000,
                 last_point_lon: 180000000,
-                scanning_mode: ScanningMode(0b11000000),
             },
+            param_set::ScanningMode(0b11000000),
             vec![(-90.0, 179.76001), (-90.0, 179.52002), (-90.0, 179.28003)],
-            vec![(90.0, 180.48), (90.0, 180.24), (90.0, 180.0)]
+            vec![(90.0, -179.52), (90.0, -179.76), (90.0, -180.0)]
         ),
     }
 
@@ -263,14 +192,22 @@ mod tests {
         ),)*) => ($(
             #[test]
             fn $name() {
-                let grid = LatLonGridDefinition {
-                    ni: 1,
-                    nj: 1,
-                    first_point_lat: $first_point_lat,
-                    first_point_lon: $first_point_lon,
-                    last_point_lat: $last_point_lat,
-                    last_point_lon: $last_point_lon,
-                    scanning_mode: ScanningMode($scanning_mode),
+                let grid = param_set::LatLonGrid {
+                    grid: param_set::Grid {
+                        ni: 1,
+                        nj: 1,
+                        initial_production_domain_basic_angle: 0,
+                        basic_angle_subdivisions: 0xffffffff,
+                        first_point_lat: $first_point_lat,
+                        first_point_lon: $first_point_lon,
+                        resolution_and_component_flags:
+                            param_set::ResolutionAndComponentFlags(0b00110000),
+                        last_point_lat: $last_point_lat,
+                        last_point_lon: $last_point_lon,
+                    },
+                    i_direction_inc: 0xffffffff,
+                    j_direction_inc: 0xffffffff,
+                    scanning_mode: param_set::ScanningMode($scanning_mode),
                 };
                 assert_eq!(grid.is_consistent_for_j(), $expected_for_j);
             }